+++ rust/sglang-server/src/api_server/layers.rs
@@ -1,14 +1,14 @@
 //! Cross-cutting tower layers shared by both transports. Each is generic over
 //! `http::Request<B>` / `http::Response<RB>`, so one implementation applies to
-//! the hand-built HTTP service and tonic's `Server::builder().layer(...)`
-//! alike — the transports differ only in their body types (see
-//! [`RejectionBody`]).
+//! the axum router and tonic's `Server::builder().layer(...)` alike — the
+//! transports differ only in their body types (see [`RejectionBody`]).
 
 pub(crate) mod access_log;
 pub(crate) mod auth;
 
-/// The peer address, inserted as a request extension by the HTTP accept loop
-/// (tonic's transport records its own `TcpConnectInfo` instead).
+/// The peer address, stamped as a request extension from axum's `ConnectInfo`
+/// in `http::app::serve` (tonic's transport records its own `TcpConnectInfo`
+/// instead).
 #[derive(Clone, Copy)]
 pub(crate) struct Peer(pub(crate) std::net::SocketAddr);
